migration: add qatzip option and CONFIG_QATZIP in code
In the previous commit, if qatzip is not installed, configure
will failed.
In this commit, We add the qatzip option in the meson_options.txt
and meson-buildoptons.sh. if you installed qatzip and would like to
In this commit, We add the qatzip option in the meson_options.txt
and meson-buildoptons.sh. if you installed qatzip and would like to
use the qatzip, We can use the "configure --enable-qatzip", it will
check the qatzip availablility from the pkg-config list. since the
qatzip .pc file is in the /usr/local/lib/pkg-config/, please "export
PKG_CONFIG_PATH=/usr/local/lib/pkg-config" and make sure the qatzip
can be shown "pkg-config --list-all|grep qatzip". One compile option
CONFIG_QATZIP is exposed and qatzip code only compiled if CONFIG_QATZIP
defined.
If use "configure" or "configure --disable-qatzip",qatzip will not
checked and the qatzip code will not compiled.

Signed-off-by: dennis.wu <dennis.wu@intel.com>